iptables_setup: selectively disable conntrack
We don't want to pollute conntrack tables with packets to/from the
instance network, or gre packets (since they might be distributed anyway
over different endpoints). So we disable conntracking for them.

This patch also moves standard setup to be done only if the setup script
is called without an explicit config file, so that the conntracking
rulers can be applied separately at each call.

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Michael Hanselmann <hansmi@google.com>