ganeti-nld: fix initial startup
When we start nld we want to perform the queries immediately. Calling
Update* is broken, though, because each Update function will think it's
being called as a result of a timer action, and thus disable the current
timer and enable a new one, which schedules one timer per update
function.

In order to solve this we just call _EnableTimers, but pass in a special
parameter to make sure the timer is set to 0 for the first run.

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Michael Hanselmann <hansmi@google.com>