ganeti-nld: update MC list periodically
Every once in a while (120s is ok, we don't expect lots of MCs to be
changing most of the time, and we're safe as long as we have some
overlap) update the MC list, making sure we can still talk to the
cluster even if the MCs change. In the future we should also save the
startup file, but just on non-nodes.

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Michael Hanselmann <hansmi@google.com>