Don't force spurious table entries in the config
Currently if a config file has no table or interface, the default
mapping is pushed to the table, even if there is an association in a
further file. Fix this by pushing an association only if a config file
has either a table or an interface (and using defaults, if not both, to
be compatible with what we did before). The default mapping, at the end,
is only pushed if no other associations are found.

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>