Fix path mtu discovery from the instances
When policy routing is used, we want icmp fragmentation-needed packets
to go directly to the instances via their direct route, rather than be
routed according to our main routing table. This iptables rule sets a
local netfilter mark on the packets that forces them to be looked up
against the separate routing table.

This initial fix requires the instance networks to be non overlapping.

Since only the fragmentation needed packets are fwmarked any other
communication from the node to the instance and vice versa is still
forbidden, through the direct path.

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Balazs Lecz <leczb@google.com>