run-in-tempdir: use system ganeti if available
If the ganeti is installed on the local system, we don't need to copy
its libraries to the tmpdir, because it will be picked up.

On the other hand, if we need to install our copy, we also need to build
it, or it won't be working (missing _autoconf.py).

This fixes "make check".

Signed-off-by: Guido Trotter <ultrotter@google.com>
Reviewed-by: Balazs Lecz <leczb@google.com>